+++ tests/test_sale_page.py
@@ -1,3 +1,4 @@
+import pytest
 from selene import browser, be, have
 from pages.locators import SalePageLocators, BaseLocators
 
@@ -49,39 +50,3 @@ def test_011_001_001_sale_breadcrumbs_is_correct(driver):
     assert expected_list_breadcrumbs == actual_list_breadcrumbs
 
 
-def test_011_011_001_women_sale_breadcrumbs_is_correct(driver):
-    # assert error !!! 'Sale' is missing
-    driver.get(SalePageLocators.LINK_WOMEN_SALE)
-    list_br = driver.find_elements(*BaseLocators.BREADCRUMBS_LIST)
-    expected_list_breadcrumbs = ['Home', 'Sale', 'Women Sale']
-    actual_list_breadcrumbs = []
-    for elem in list_br:
-        actual_list_breadcrumbs.append(elem.text)
-    assert expected_list_breadcrumbs == actual_list_breadcrumbs
-
-
-def test_011_011_002_breadcrumbs_redirection_from_women_sale(driver):
-    driver.get(SalePageLocators.LINK_WOMEN_SALE)
-    breadcrumbs_links = browser.find_elements(*BaseLocators.BREADCRUMBS_LINKS)
-    links = [elem.get_attribute('href') for elem in breadcrumbs_links]
-    for elem in links:
-        browser.get(elem)
-        assert browser.current_url in SalePageLocators.BREADCRUMBS_LINKS_ON_PAGE_WOMEN_SALE
-
-
-def test_011_016_001_women_tees_breadcrumbs_is_correct(driver):
-    driver.get(SalePageLocators.LINK_TEES_WOMEN)
-    list_br = browser.find_elements(*BaseLocators.BREADCRUMBS_LIST)
-    expected_list_breadcrumbs = ['Home', 'Women', 'Tops', 'Tees']
-    actual_list_breadcrumbs = []
-    for elem in list_br:
-        actual_list_breadcrumbs.append(elem.text)
-    assert expected_list_breadcrumbs == actual_list_breadcrumbs
-
-
-def test_011_016_002_breadcrumbs_redirection_from_women_tees(driver):
-    # check if all link are there
-    driver.get(SalePageLocators.LINK_TEES_WOMEN)
-    breadcrumb_links = browser.find_elements(*BaseLocators.BREADCRUMBS_LINKS)
-    links = [elem.get_attribute('href') for elem in breadcrumb_links]
-    assert links == SalePageLocators.BREADCRUMBS_LINKS_ON_PAGE_TEES
